[SPARK-26990][SQL][BACKPORT-2.4] FileIndex: use user specified field names if possible
## What changes were proposed in this pull request?

Back-port of #23894 to branch-2.4.

WIth the following file structure:
```
/tmp/data
└── a=5
```

In the previous release:
```
scala> spark.read.schema("A int, ID long").parquet("/tmp/data/").printSchema
root
 |-- ID: long (nullable = true)
 |-- A: integer (nullable = true)
```

While in current code:
```
scala> spark.read.schema("A int, ID long").parquet("/tmp/data/").printSchema
root
 |-- ID: long (nullable = true)
 |-- a: integer (nullable = true)
```

We can see that the partition column name `a` is different from `A` as user specifed. This PR is to fix the case and make it more user-friendly.

Closes #23894 from gengliangwang/fileIndexSchema.

Authored-by: Gengliang Wang <gengliang.wangdatabricks.com>
Signed-off-by: Wenchen Fan <wenchendatabricks.com>

## How was this patch tested?

Unit test

Closes #23909 from bersprockets/backport-SPARK-26990.

Authored-by: Gengliang Wang <gengliang.wang@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>