[SPARK-30312][SQL][2.4] Preserve path permission and acl when truncate table
This patch proposes to preserve existing permission/acls of paths when truncate table/partition.
Note that this is backport of #26956 to branch-2.4.

When Spark SQL truncates table, it deletes the paths of table/partitions, then re-create new ones. If permission/acls were set on the paths, the existing permission/acls will be deleted.

We should preserve the permission/acls if possible.

Yes. When truncate table/partition, Spark will keep permission/acls of paths.

Unit test and manual test as shown in #26956.

Closes #27173 from viirya/truncate-table-permission-2.4.

Authored-by: Liang-Chi Hsieh <liangchi@uber.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>